Set "Electrum Server is out of sync" delay to 30 days
This setting is a GUI setting (the wallet will check and disconnect servers whose tip
is too old), which can be safely set to 30 days, which will allow users to use their
wallet as soon as we're connected and check our electrum server.
In the very rare instances where users were able to create and try to publish a tx while
the wallet was syncing AND their UTXO became invalid (double-spent, or spent by them from
another wallet initialized with the same seed) then publishing the tx will fail and after
a few seconds the wallet will be updated.